two points after decimal point & bugs

+++ myBank.h
@@ -1,6 +1,6 @@
-void openAccount (int amount);
+void openAccount (double amount);
 void ballance (int account_number);
-void deposit(int account_number, int amount);
+void deposit(int account_number, double amount);
 void withdrawal(int account_number, double amount);
 void closeAccount(int account_number);
 void interest( double interest_rate);
